Dedupe package.json scripts using env. vars
This should make it easier to change the scripts in the future, without
having to change them in multiple places. It also makes it easier for
developers to run various scripts after choosing a browser and setting
the `target` environment variable. For example:

```bash
export target=firefox
yarn watch &
yarn launch
```

[cross-env] is used to maintain Windows compatibility

[cross-env]: https://github.com/kentcdodds/cross-env

---

Also replace `npm run` with `yarn`, which is easier to type, and matches
CONTRIBUTING.md.